backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
httpx==0.25.2
gspread==5.12.0
google-auth==2.25.2
pydantic==2.5.2
